bug parcial checkToken con token viejo

+++ src/app/auth-guard.guard.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
 import { LoginService } from './service/login.service';
 import { PatronGuard } from './patron.guard';
+import { AppComponent } from './app.component';
 
 @Injectable({
   providedIn: 'root'
@@ -10,41 +11,32 @@ export class AuthGuardGuard implements CanActivate {
   constructor(public login: LoginService, public router: Router, public patronGuard: PatronGuard) {
 
   }
-  // async canActivate() {
-  //   //
-  //     console.log("authGuard" + this.router.url);
-  //     var resp: any = false;
-  //     this.login.checkToken("api/checkToken", { token: localStorage.getItem("token") }).then(async (data) => {
-  //     resp=true;
-  //     console.log("llego hasta aca" + this.router.url + " " + resp);
-  //     return (resp);
-  //   })
-  //     .catch(data => {
-  //         localStorage.removeItem("token");
-  //         this.router.navigateByUrl("welcome");
-  //         resp = false;
-  //         return(resp);
-  //   });
-  // }
+
   async canActivate(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot,
   ) {
     //
+    console.error("authGuard");
     console.log("authGuard" + this.router.url);
     var resp: any = false;
-    await this.login.checkToken("api/checkToken", { token: localStorage.getItem("token") }).then(data => {
-      resp = true;
-      console.log("llego hasta aca" + this.router.url + " " + resp);
-      localStorage.removeItem("onboarding");
+    console.log(AppComponent.CheckedToken);
+    if (await AppComponent.CheckedToken == true) {
       return true;
-    }).catch(data => {
-      localStorage.removeItem("token");
-      this.router.navigateByUrl("welcome");
-      localStorage.clear();
-      resp = false;
-      return false
-    });
+    }
+    else
+      await this.login.checkToken("api/checkToken", { token: localStorage.getItem("token") }).then(data => {
+        resp = true;
+        console.log("llego hasta aca" + this.router.url + " " + resp);
+        localStorage.removeItem("onboarding");
+        return true;
+      }).catch(data => {
+        localStorage.removeItem("token");
+        this.router.navigateByUrl("welcome");
+        localStorage.clear();
+        resp = false;
+        return false
+      });
     return resp;
   }
 }
